server/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
python-dotenv
google-genai
pypdf
python-docx
jsonschema
